+++ apps/dashboard/src/app/app-routing.module.ts
@@ -141,6 +141,10 @@ const routes: Routes = [
                   ),
               },
               {
+                path: 'synonyms',
+                loadChildren: () => import('./synonyms/synonyms.module').then((m) => m.SynonymsModule),
+              },
+              {
                 path: 'manage',
                 component: KnowledgeBoxProfileComponent,
               },