+++ libs/common/src/lib/navbar/navbar.component.ts
@@ -1,6 +1,6 @@
 import { Component, Input, OnDestroy, OnInit } from '@angular/core';
 import { animate, style, transition, trigger } from '@angular/animations';
-import { map, merge, of, shareReplay, Subject } from 'rxjs';
+import { combineLatest, map, merge, Observable, of, shareReplay, Subject } from 'rxjs';
 import { filter, takeUntil } from 'rxjs/operators';
 import { AppService, NavigationService } from '../services';
 import { SDKService, StateService, STFTrackingService } from '@flaps/core';
@@ -33,13 +33,6 @@ export class NavbarComponent extends SmallNavbarDirective implements OnInit, OnD
     ),
   );
   kbUrl: string = '';
-  isActivityEnabled = this.tracking.isFeatureEnabled('view-activity').pipe(shareReplay(1));
-  isEntitiesEnabled = this.tracking.isFeatureEnabled('manage-entities').pipe(shareReplay(1));
-  isUsersEnabled = this.tracking.isFeatureEnabled('manage-users').pipe(shareReplay(1));
-  isWidgetsEnabled = this.tracking.isFeatureEnabled('manage-widgets').pipe(shareReplay(1));
-  isAPIKeysEnabled = this.tracking.isFeatureEnabled('manage-api-keys').pipe(shareReplay(1));
-  isBillingEnabled = this.tracking.isFeatureEnabled('billing').pipe(shareReplay(1));
-  isTrainingEnabled = this.tracking.isFeatureEnabled('training').pipe(shareReplay(1));
 
   isAdminOrContrib = this.sdk.currentKb.pipe(map((kb) => !!kb.admin || !!kb.contrib));
   isAdmin = this.sdk.currentKb.pipe(map((kb) => !!kb.admin));
@@ -47,6 +40,21 @@ export class NavbarComponent extends SmallNavbarDirective implements OnInit, OnD
   kb = this.sdk.currentKb;
   accountUrl = this.account.pipe(map((account) => this.navigationService.getAccountManageUrl(account!.slug)));
   isAccountManager = this.account.pipe(map((account) => account!.can_manage_account));
+
+  isActivityEnabled = this.tracking.isFeatureEnabled('view-activity').pipe(shareReplay(1));
+  isEntitiesEnabled = this.tracking.isFeatureEnabled('manage-entities').pipe(shareReplay(1));
+  isUsersEnabled = this.tracking.isFeatureEnabled('manage-users').pipe(shareReplay(1));
+  isWidgetsEnabled = this.tracking.isFeatureEnabled('manage-widgets').pipe(shareReplay(1));
+  isAPIKeysEnabled = this.tracking.isFeatureEnabled('manage-api-keys').pipe(shareReplay(1));
+  isBillingEnabled = this.tracking.isFeatureEnabled('billing').pipe(shareReplay(1));
+  isTrainingEnabled = this.tracking.isFeatureEnabled('training').pipe(shareReplay(1));
+  isSynonymsEnabled: Observable<boolean> = combineLatest([
+    this.tracking.isFeatureEnabled('manage-synonyms').pipe(shareReplay(1)),
+    this.account.pipe(
+      filter((account) => !!account),
+      map((account) => account?.type),
+    ),
+  ]).pipe(map(([featureEnabled, accountType]) => featureEnabled && accountType === 'stash-business'));
 
   standalone = this.sdk.nuclia.options.standalone;