+++ libs/sdk-core/src/lib/db/db.models.ts
@@ -1,4 +1,10 @@
-export type AccountTypes = 'stash-basic' | 'stash-team' | 'stash-enterprise';
+export type AccountTypes =
+  | 'stash-basic'
+  | 'stash-team'
+  | 'stash-enterprise'
+  | 'stash-pro'
+  | 'stash-growth'
+  | 'stash-business';
 
 export const NUA_KEY = 'NUCLIA_NUA_KEY';
 export const NUA_CLIENT = 'NUCLIA_NUA_CLIENT';